+++ src/actions/user.js
@@ -16,36 +16,3 @@ export function fetchUserDetails(userId) {
     .catch(response => console.log('RESPONSE-error: ', response))
   }
 }
-
-export function checkLoginUser() {  //First thing on app load - check if the user is logged in.
-  return (dispatch) => {
-    Panoptes.auth.checkCurrent()
-      .then((user) => {
-        dispatch(setLoginUser(user));
-      });
-  }
-}
-
-export function setLoginUser(user) {
-  return (dispatch) => {
-    dispatch({
-      type: types.SET_LOGIN_USER,
-      user
-    });
-  };
-}
-
-export function loginToPanoptes() {  //Returns a login page URL for the user to navigate to.
-  return (dispatch) => {
-    return Panoptes.oauth.signIn(window.location.host)
-  }
-}
-
-export function logoutFromPanoptes() {
-  return (dispatch) => {
-    Panoptes.oauth.signOut()
-      .then(user => {
-        dispatch(setLoginUser(user));
-      });
-  }
-}
