+++ src/constants/config.js
@@ -1,3 +1,3 @@
-const appId = {
+const config = {
   panoptesAppId: '2abed6ac014d4a8416a1f578de922600451a10c759dfeab6c9f0edbeb91c8f88',
 };
