+++ src/reducers/user.js
@@ -20,3 +20,4 @@ export function user(state = initialState, action) {
       return state;
   }
 }
+
